+++ example/lib/app_model.dart
@@ -1,14 +1,20 @@
 import 'package:flutter/material.dart';
 
-abstract class AppModel extends ChangeNotifier
-{
-    void incrementCounter();
+import 'main.dart';
 
-    int get counter;
+abstract class AppModel extends ChangeNotifier {
+  void incrementCounter();
+
+  int get counter;
 }
 
-class AppModelImplementation   extends AppModel  {
+class AppModelImplementation extends AppModel {
   int _counter = 0;
+
+  AppModelImplementation() {
+    /// lets pretend we have to do some async initilization
+    Future.delayed(Duration(seconds: 3)).then((_) => getIt.signalReady(this));
+  }
 
   @override
   int get counter => _counter;
@@ -18,10 +24,4 @@ class AppModelImplementation   extends AppModel  {
     _counter++;
     notifyListeners();
   }
-  
 }
-
-
-
-
-